5.5 Приводит файловую структуру в порядок
