+++ src/entities/product/product-card/index.js
