+++ src/entities/product/product-card/style.pcss
