+++ src/features/cart/add-to-cart.js
@@ -1,4 +1,4 @@
-import { addToCart } from "../../entities/product/model";
+import { addToCart } from "../../shared/store/actions";
 
 export const addProductToCart = async (productId) => {
   await addToCart(productId);
