+++ src/pages/catalog/index.js
@@ -1,46 +1,9 @@
+import { categories } from "../../shared/constants/common";
 import { getMeta, getPage } from "../../shared/lib";
 import { CatalogList } from "../../widgets/catalog-list/ui";
 import { Filter } from "../../widgets/filter/filter";
 import { Footer } from "../../widgets/footer/ui/footer";
 import { Header } from "../../widgets/header/ui/header";
-
-const categories = {
-  all: {
-    name: "Все курсы",
-    searchParam: "all",
-    isChecked: true,
-  },
-  backend: {
-    name: "Бэкенд-разработка",
-    searchParam: "backend",
-    isChecked: false,
-  },
-  web: {
-    name: "Веб-разработка",
-    searchParam: "web",
-    isChecked: false,
-  },
-  mobile: {
-    name: "Мобильная разработка",
-    searchParam: "mobile",
-    isChecked: false,
-  },
-  data: {
-    name: "Анализ данных",
-    searchParam: "data",
-    isChecked: false,
-  },
-  structure: {
-    name: "IT-инфраструктура",
-    searchParam: "structure",
-    isChecked: false,
-  },
-  management: {
-    name: "Управление разработкой",
-    searchParam: "management",
-    isChecked: false,
-  },
-};
 
 export default () => {
   return getPage({
