+++ src/shared/constants/common.js
@@ -1,0 +1,37 @@
+export const categories = {
+  all: {
+    name: "Все курсы",
+    searchParam: "all",
+    isChecked: true,
+  },
+  backend: {
+    name: "Бэкенд-разработка",
+    searchParam: "backend",
+    isChecked: false,
+  },
+  web: {
+    name: "Веб-разработка",
+    searchParam: "web",
+    isChecked: false,
+  },
+  mobile: {
+    name: "Мобильная разработка",
+    searchParam: "mobile",
+    isChecked: false,
+  },
+  data: {
+    name: "Анализ данных",
+    searchParam: "data",
+    isChecked: false,
+  },
+  structure: {
+    name: "IT-инфраструктура",
+    searchParam: "structure",
+    isChecked: false,
+  },
+  management: {
+    name: "Управление разработкой",
+    searchParam: "management",
+    isChecked: false,
+  },
+};
