+++ src/widgets/filter/ui/style.pcss
