+++ src/widgets/footer/ui/index.js
