+++ src/widgets/slider/ui/style.pcss
